+++ ouds_accessibility_plugin/lib/ouds_accessibility_plugin.dart
@@ -13,12 +13,14 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
-
 import 'ouds_accessibility_plugin_platform_interface.dart';
 
 class OudsAccessibilityPlugin {
+
+
   // Singleton for accessing a specific platform
-  static OudsAccessibilityPluginPlatform get _platform => OudsAccessibilityPluginPlatform.instance;
+  static OudsAccessibilityPluginPlatform get _platform =>
+      OudsAccessibilityPluginPlatform.instance;
 
   /// Returns whether the High Contrast mode is enabled.
   static Future<bool> isHighContrastEnabled(BuildContext context) async {