+++ ouds_core/lib/components/button/internal/ouds_button_border_modifier.dart
@@ -49,93 +49,64 @@ class OudsButtonBorderModifier {
   static BorderSide _getFocusedBorderColor(BuildContext context, OudsButtonAppearance appearance) {
     final theme = OudsTheme.of(context);
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
-    switch (appearance) {
-      case OudsButtonAppearance.strong:
-        return BorderSide.none;
-      case OudsButtonAppearance.brand:
-        return BorderSide.none;
-      case OudsButtonAppearance.minimal:
-        return BorderSide.none;
-      case OudsButtonAppearance.negative:
-        return BorderSide.none;
-      default:
-        return BorderSide(
-          color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultFocus : theme.componentsTokens(context).button.colorBorderDefaultFocus,
-          width: onColoredSurface ? theme.componentsTokens(context).button.borderWidthDefaultInteractionMono : theme.componentsTokens(context).button.borderWidthDefaultInteraction,
-        );
+    if (appearance == OudsButtonAppearance.defaultAppearance) {
+      return BorderSide(
+        color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultFocus : theme.componentsTokens(context).button.colorBorderDefaultFocus,
+        width: onColoredSurface ? theme.componentsTokens(context).button.borderWidthDefaultInteractionMono : theme.componentsTokens(context).button.borderWidthDefaultInteraction,
+      );
+    } else {
+      return BorderSide.none;
     }
   }
 
   static BorderSide _getEnabledBorderColor(BuildContext context, OudsButtonAppearance appearance) {
     final theme = OudsTheme.of(context);
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
-    switch (appearance) {
-      case OudsButtonAppearance.strong:
-        return BorderSide.none;
-      case OudsButtonAppearance.brand:
-        return BorderSide.none;
-      case OudsButtonAppearance.minimal:
-        return BorderSide.none;
-      case OudsButtonAppearance.negative:
-        return BorderSide.none;
-      default:
-        return BorderSide(
-            color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultEnabled : theme.componentsTokens(context).button.colorBorderDefaultEnabled, width: theme.componentsTokens(context).button.borderWidthDefault);
+    if (appearance == OudsButtonAppearance.defaultAppearance) {
+      return BorderSide(
+          color: onColoredSurface
+              ? theme.componentsTokens(context).buttonMono.colorBorderDefaultEnabled
+              : theme.componentsTokens(context).button.colorBorderDefaultEnabled,
+          width: theme.componentsTokens(context).button.borderWidthDefault);
+   } else {
+      return BorderSide.none;
     }
   }
 
   static BorderSide _getHoverBorderColor(BuildContext context, OudsButtonAppearance appearance) {
     final theme = OudsTheme.of(context);
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
-    switch (appearance) {
-      case OudsButtonAppearance.strong:
-        return BorderSide.none;
-      case OudsButtonAppearance.brand:
-        return BorderSide.none;
-      case OudsButtonAppearance.minimal:
-        return BorderSide.none;
-      case OudsButtonAppearance.negative:
-        return BorderSide.none;
-      default:
-        return BorderSide(
-            color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultEnabled : theme.componentsTokens(context).button.colorBorderDefaultHover, width: theme.componentsTokens(context).button.borderWidthDefaultInteraction);
+    if (appearance == OudsButtonAppearance.defaultAppearance) {
+      return BorderSide(
+          color: onColoredSurface
+              ? theme.componentsTokens(context).buttonMono.colorBorderDefaultEnabled
+              : theme.componentsTokens(context).button.colorBorderDefaultHover,
+          width: theme.componentsTokens(context).button.borderWidthDefaultInteraction);
+    } else {
+      return BorderSide.none;
     }
   }
 
   static BorderSide _getPressedBorderColor(BuildContext context, OudsButtonAppearance appearance) {
     final theme = OudsTheme.of(context);
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
-    switch (appearance) {
-      case OudsButtonAppearance.strong:
-        return BorderSide.none;
-      case OudsButtonAppearance.brand:
-        return BorderSide.none;
-      case OudsButtonAppearance.minimal:
-        return BorderSide.none;
-      case OudsButtonAppearance.negative:
-        return BorderSide.none;
-      default:
-        return BorderSide(
-            color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultPressed : theme.componentsTokens(context).button.colorBorderDefaultPressed,
-            width: theme.componentsTokens(context).button.borderWidthDefaultInteraction);
+    if (appearance == OudsButtonAppearance.defaultAppearance) {
+      return BorderSide(
+          color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultPressed : theme.componentsTokens(context).button.colorBorderDefaultPressed,
+          width: theme.componentsTokens(context).button.borderWidthDefaultInteraction);
+    } else {
+      return BorderSide.none;
     }
   }
 
   static BorderSide _getDisabledBorderColor(BuildContext context, OudsButtonAppearance appearance) {
     final theme = OudsTheme.of(context);
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
-    switch (appearance) {
-      case OudsButtonAppearance.strong:
-        return BorderSide.none;
-      case OudsButtonAppearance.brand:
-        return BorderSide.none;
-      case OudsButtonAppearance.minimal:
-        return BorderSide.none;
-      case OudsButtonAppearance.negative:
-        return BorderSide.none;
-      default:
-        return BorderSide(
-            color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultDisabled : theme.componentsTokens(context).button.colorBorderDefaultDisabled, width: theme.componentsTokens(context).button.borderWidthDefault);
+    if (appearance == OudsButtonAppearance.defaultAppearance) {
+      return BorderSide(
+          color: onColoredSurface ? theme.componentsTokens(context).buttonMono.colorBorderDefaultDisabled : theme.componentsTokens(context).button.colorBorderDefaultDisabled, width: theme.componentsTokens(context).button.borderWidthDefault);
+    } else {
+      return BorderSide.none;
     }
   }
 
@@ -157,9 +128,9 @@ class OudsButtonBorderModifier {
     final buttonRounded = OudsThemeConfigModel.of(context)?.button?.rounded ?? false;
     switch (buttonRounded) {
       case true:
-        return button.borderRadiusRounded;
+        return button.borderRadiusRounded + OudsTheme.of(context).borderTokens.widthFocus;
       case false:
-        return button.borderRadiusDefault;
+        return button.borderRadiusDefault ;
     }
   }
 }